Fix for ErrorException: Notice: A non well formed numeric value encountered.
ErrorException: Notice: A non well formed numeric value encountered
  File "/vendor/squeaky-minds/squeaky-minds-php-helper/src/SqueakyMindsPhpHelper.php", line 46, in SqueakyMindsPhpHelper::uuid
    return (string)md5(uniqid(rand() + MicroTime(), 1));
  File "/app/_includes/_autoload/utilityContainer.inc.php", line 47, in utilityContainer::setNoncesUUID
    $_SESSION["NoncesUUID"] = SqueakyMindsPhpHelper::uuid();
  File "/index.php", line 110
    utilityContainer::setNoncesUUID();

+++ src/SqueakyMindsPhpHelper.php
@@ -43,7 +43,7 @@ class SqueakyMindsPhpHelper
      */
     static public function uuid() 
     {
-        return (string)md5(uniqid(rand() + MicroTime(), 1));
+        return (string)md5(uniqid((int)rand() + (int)MicroTime(), 1));
     }
